MatchingGame
Created Matching Game, this includes a functioning system where cards are checked to see if matched, if matched add match count. This count is kept until user loses. If user loses their score is kept in statisitcs view. Also have difficulties with animated text and a gif for the main page, also tried to make it christmas themed to the best of my ability.